[Chore] 자료형 변경
자료형 변경 및 플레이어 회전 로직 일부 수정

+++ Assets/WorkSpace/YJH/Scripts/EnergyBolt.cs
@@ -8,7 +8,7 @@ public class EnergyBolt : MonoBehaviour, IAttackType, IDamager
     [SerializeField] float speed=5;
     [SerializeField] float maxRange=5;
     [SerializeField] int energyCost=1;
-    [SerializeField] float damage;//float으로?
+    [SerializeField] int damage;//float으로?
     [SerializeField] bool canPierce=false;
     [SerializeField] Transform firePos;
     [SerializeField] GameObject hitEffect;
@@ -30,7 +30,7 @@ public class EnergyBolt : MonoBehaviour, IAttackType, IDamager
     }
     public float MaxRange { get { return maxRange; } set { maxRange = value; } }
     public int EnergyCost { get { return energyCost; } set { energyCost = value; } }
-    public float Damage { get { return damage; } set { damage = value; } }//float으로?
+    public int Damage { get { return damage; } set { damage = value; } }//float으로?
     public bool CanPierce { get { return canPierce; }set { canPierce = value; } }
     public Transform FirePos { get { return firePos; }set { firePos = value; } }
     public GameObject HitEffect { get { return hitEffect; }set { hitEffect = value; } }
